Add .DS_Store to .gitignore, rename Moscowvitz to ClipVitz, rename folder to ClipVitz-backend
