+++ clipvitz-backend/server.js
@@ -1,11 +1,10 @@
-
-   const express = require('express');
+const express = require('express');
    const cors = require('cors');
-   const Moscowvitz = require('./Moscowvitz');
+   const ClipVitz = require('./ClipVitz');
 
    const app = express();
    const port = process.env.PORT || 3000;
-   const moscowvitz = new Moscowvitz();
+   const clipVitz = new ClipVitz();
 
    app.use(cors());
    app.use(express.json());
@@ -17,7 +16,7 @@
          return res.status(400).json({ error: 'Prompt is required' });
        }
        console.log("Generating social post for prompt:", prompt);
-       const post = await moscowvitz.generateSocialPost(prompt);
+       const post = await clipVitz.generateSocialPost(prompt);
        res.json({ story: post });
      } catch (error) {
        console.error("Error generating social post:", error);
